fixed some bugs, added some mutex's and error handling, need to work on ik in fk scene, tip position relative to parent handle isnt favourable, should use world coords and subtract relative to base

+++ programming/client/kinematics.h
@@ -9,6 +9,6 @@ void fk_mod(move* move_ptr);
 void ik_RRR_arm(move* move_ptr, char* plane);
 
 void inverse_kinematics(move* move_ptr, info* info_ptr, float* position, double* angles);
-void control_kinematics(info* info_ptr, move* move_ptr, int x, int y, int z);
+void control_kinematics(info* info_ptr, move* move_ptr, float x, float y, float z);
 
 #endif // !KINEMATICS.H
